requirements.txt
discord
chess
python-chess
Pillow
#io
firebase_admin
py-dotenv